Massive renderer upgrade to use 32-bit integer math in order to avoid savage overflow problems that have plagued the DOOM engine since its conception. Also, some of joe's new Small functions.

+++ eternity/source/r_bsp.c
@@ -229,7 +229,9 @@ void R_ClearClipSegs(void)
 }
 
 #ifdef R_PORTALS
-boolean R_SetupPortalClipsegs(short *top, short *bottom)
+// haleyjd: DEBUG
+//boolean R_SetupPortalClipsegs(short *top, short *bottom)
+boolean R_SetupPortalClipsegs(int *top, int *bottom)
 {
    int i = 0;
    cliprange_t *seg;
